rename Controllers name && rename RefreshToken model name to refresh_tokens
