+++ routes/api/index.js
@@ -1,36 +1,36 @@
 const express = require("express");
 const router = express.Router();
 
-const postController = require("../../controllers/p_ostController");
-const userController = require("../../controllers/u_serController");
-const AuthController = require("../../controllers/A_uthController");
+const PostController = require("../../controllers/PostController");
+const UserController = require("../../controllers/UserController");
+const AuthController = require("../../controllers/AuthController");
 
 //Post route
-router.post("/post", postController.apiCreatePost);
+router.post("/post", Post.apiCreatePost);
 
-router.get("/post", postController.apiGetAllpost);
-router.get("/post/:title", postController.apiSearchAllpost);
-router.delete("/post=:postId", postController.deletePost);
+router.get("/post", PostController.apiGetAllpost);
+router.get("/post/:title", PostController.apiSearchAllpost);
+router.delete("/post=:postId", PostController.deletePost);
 
 //User route
-router.get("/user", userController.apiGetAlluser);
-router.get("/user=:id", userController.apiGetOneuser);
+router.get("/user", UserController.apiGetAlluser);
+router.get("/user=:id", UserController.apiGetOneuser);
 
-router.post("/user/register", userController.apiCreateNewUser);
+router.post("/user/register", UserController.apiCreateNewUser);
 
 //Follower route
-router.post("/user/follow", userController.apiFollowUser);
+router.post("/user/follow", UserController.apiFollowUser);
 
 //Auth
-router.post("/user/login", userController.apiLoginUser);
+router.post("/user/login", UserController.apiLoginUser);
 router.post("/user/token", AuthController.renewAccessToken);
 router.delete("/user/logout", AuthController.deleteToken);
 
-router.patch("/user/updatePassword", userController.apiUpdateUserPassword);
+router.patch("/user/updatePassword", UserController.apiUpdateUserPassword);
 router.patch(
   "/user/updatePhone",
   AuthController.authenticateToken,
-  userController.apiUpdateUserPhone
+  UserController.apiUpdateUserPhone
 );
 
 module.exports = router;
